Game Objectのキャッシュを実装
